sequentialize/mutex vs sequentialize/actor
We now provide two ways to sequentialize access to a function:
sequentialize/actor and sequentialize/mutex.
The former runs the function in its own actor, with its own dynamic
context. The latter runs the function in the dynamic context of the
caller. The function sequentialize can be mutated to be one or the
other. It used to always be sequentialize/actor. Now it defaults
to sequentialize/mutex for the increased performance.