Fix subtle bug in git-fu
Use nested let*'s and if's instead of def's in relevant functions (that were
just modified that way), because when compiled (as opposed to interpreted),
these def's will be transformed to letrec's that will reorder the evaluation
compared to the source code, leading to some elements being used before they
are defined. Therefore the code worked fine when interpreted when ./build.ss
was evaluated the first time around, but segfaulted after it was compiled.